FEAT: Notification hook working
