+++ frontend/src/components/useNotifications/useNotifications.tsx
@@ -1,39 +1,13 @@
-import React, { useState } from "react";
-import Snackbar from "@mui/material/Snackbar";
-import type { SnackbarProps } from "@mui/material/Snackbar";
+import { useContext } from "react";
 import { NotificationsContext } from "./NotificationsContext";
 
-export interface NotificationsProviderProps {
-  children?: React.ReactNode;
-  notificationProps?: SnackbarProps;
-}
+export const useNotifications = () => {
+  const context = useContext(NotificationsContext);
+  if (!context) {
+    throw new Error(
+      "useNotifications must be used within a NotificationsProvider",
+    );
+  }
 
-const NotificationsProvider = ({
-  children,
-  notificationProps,
-}: NotificationsProviderProps) => {
-  const [message, setMessage] = useState("");
-  const [open, setOpen] = useState(false);
-
-  const showNotification = (message: string) => {
-    setMessage(message);
-    setOpen(true);
-  };
-
-  return (
-    <NotificationsContext.Provider value={{ showNotification }}>
-      {children}
-
-      <Snackbar
-        anchorOrigin={{ vertical: "top", horizontal: "center" }}
-        message={message}
-        open={open}
-        autoHideDuration={3000}
-        onClose={() => setOpen(false)}
-        {...notificationProps}
-      />
-    </NotificationsContext.Provider>
-  );
+  return context;
 };
-
-export { NotificationsProvider };
